Provide a safe ISR wrapper macro
We must avoid allocating a stack frame in the direct inerrupt service
routine because the `iret` will execute before the stack is cleaned up.
However, without a proper stack frame, we'll quickly corrupt the values
on the stack that are necessary for `iret()` to return from the ISR.
Thus, we create a small trampoline function which doesn't have a stack
frame but calls a locally-scoped function that contains the interrupt
handler code. This function will properly enter and leave a new stack
frame.