Fix handling of non maskable interrupts
These interrupts can occur at any time, even during an otherwise correct
instruction sequence. To avoid handling these on a userspace stack just
before executing the `sysret` instruction, we force these interrupts to
load a special, safe stack.

We reserve space for the stack in the kernel .bss section. We then
reference this space in the TSS. When the interrupt occurs, ths ist1
field of the TSS will be loaded into rsp.