Initialize fs/gs registers and reorder GDT
SeaBIOS sets fs/gs to 0x18 which becomes invalid after we change GDT.
This only causes a problem when restoring registers after interrupts.

Either of these changes would fix. But in the GDT module we define 0x18 as the
user space code segment, so it should probably be 64 bit by default.